Multi-Purpose shovel
To access Onshape in a web browser, follow these steps:

1. Open your web browser.
2. Navigate to the Onshape website (https://www.onshape.com/).
3. Create an account by clicking on the "Sign Up" or "Create Account" button.
4. Fill in the required information to register for an account.

Once you have an account, you can use the link provided to view the multipurpose shovel design. Share the specific link with others, or paste it into your browser to access the design directly. Make sure the design is shared publicly or with the intended individuals for them to view it.